move test files to e2e folder
