+++ src/integrationTest/codewhisperer/e2e/serviceInvocations.test.ts
@@ -6,19 +6,19 @@
 import * as assert from 'assert'
 import * as vscode from 'vscode'
 import * as path from 'path'
-import { setValidConnection, skiptTestIfNoValidConn } from '../util/codewhispererUtil'
-import { ConfigurationEntry } from '../../codewhisperer/models/model'
-import * as codewhispererClient from '../../codewhisperer/client/codewhisperer'
-import { RecommendationHandler } from '../../codewhisperer/service/recommendationHandler'
+import { setValidConnection, skiptTestIfNoValidConn } from '../../util/codewhispererUtil'
+import { ConfigurationEntry } from '../../../codewhisperer/models/model'
+import * as codewhispererClient from '../../../codewhisperer/client/codewhisperer'
+import { RecommendationHandler } from '../../../codewhisperer/service/recommendationHandler'
 import {
     createMockTextEditor,
     createTextDocumentChangeEvent,
     resetCodeWhispererGlobalVariables,
-} from '../../test/codewhisperer/testUtil'
-import { KeyStrokeHandler } from '../../codewhisperer/service/keyStrokeHandler'
-import { sleep } from '../../shared/utilities/timeoutUtils'
-import { invokeRecommendation } from '../../codewhisperer/commands/invokeRecommendation'
-import { getTestWorkspaceFolder } from '../integrationTestsUtilities'
+} from '../../../test/codewhisperer/testUtil'
+import { KeyStrokeHandler } from '../../../codewhisperer/service/keyStrokeHandler'
+import { sleep } from '../../../shared/utilities/timeoutUtils'
+import { invokeRecommendation } from '../../../codewhisperer/commands/invokeRecommendation'
+import { getTestWorkspaceFolder } from '../../integrationTestsUtilities'
 
 describe('CodeWhisperer service invocation', async function () {
     let validConnection: boolean
